# app/lib/api-zod/src/generated/types/autopilotBotUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * At least one field should be provided.
 */
export interface AutopilotBotUpdate {
  running?: boolean;
  capital?: number;
  drawdown?: number;
}