# app/lib/api-zod/src/generated/types/autopilotLogLine.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AutopilotLogLine {
  id: string;
  /** HH:MM:SS wall-clock time of the log line. */
  time: string;
  text: string;
}