# app/lib/api-zod/src/generated/types/autopilotPnlDay.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AutopilotPnlDay {
  /** ISO date (YYYY-MM-DD) the P&L belongs to. */
  day: string;
  /** Simulated bot P&L for that day in USD. */
  pnl: number;
}